+++ packages/core/src/test/kbjwt.spec.ts
@@ -117,6 +117,7 @@ describe('KB JWT', () => {
     const verified = await decoded.verifyKB({
       verifier: testVerifier,
       payload,
+      nonce: 'nonce',
     });
     expect(verified).toStrictEqual({
       header: {
@@ -177,7 +178,11 @@ describe('KB JWT', () => {
     const encodedKbJwt = await kbJwt.sign(testSigner);
     const decoded = KBJwt.fromKBEncode(encodedKbJwt);
     try {
-      await decoded.verifyKB({ verifier: testVerifier, payload });
+      await decoded.verifyKB({
+        verifier: testVerifier,
+        payload,
+        nonce: 'nonce',
+      });
     } catch (e: unknown) {
       const error = e as SDJWTException;
       expect(error.message).toBe('Invalid Key Binding Jwt');
@@ -222,7 +227,11 @@ describe('KB JWT', () => {
     const encodedKbJwt = await kbJwt.sign(testSigner);
     const decoded = KBJwt.fromKBEncode(encodedKbJwt);
     try {
-      await decoded.verifyKB({ verifier: testVerifier, payload });
+      await decoded.verifyKB({
+        verifier: testVerifier,
+        payload,
+        nonce: 'nonce',
+      });
     } catch (e: unknown) {
       const error = e as SDJWTException;
       expect(error.message).toBe('Verify Error: Invalid JWT Signature');
@@ -268,7 +277,11 @@ describe('KB JWT', () => {
     const decoded = KBJwt.fromKBEncode(encodedKbJwt);
     decoded.signature = undefined;
     try {
-      await decoded.verifyKB({ verifier: testVerifier, payload });
+      await decoded.verifyKB({
+        verifier: testVerifier,
+        payload,
+        nonce: 'nonce',
+      });
     } catch (e: unknown) {
       const error = e as SDJWTException;
       expect(error.message).toBe('Verify Error: Invalid JWT');
@@ -324,6 +337,7 @@ describe('KB JWT', () => {
     const verified = await decoded.verifyKB({
       verifier: testVerifier,
       payload,
+      nonce: 'nonce',
     });
     expect(verified).toStrictEqual({
       header: {